requirements.txt
python-dotenv~=0.20.0
fastapi~=0.78.0
uvicorn[standard]~=0.17.6
databases[sqlite]~=0.5.5
plotly~=5.8.0
dash~=2.4.1
dash-extendable-graph~=1.3.0
pandas~=1.4.2
requests~=2.27.1
